playwright_based_code/requirements.txt
google-api-python-client 
google-auth 
google-auth-oauthlib 
google-auth-httplib2 
pandas 
openpyxl
streamlit 
playwright
